refactor(client): extract validation to shared module (stored proc prep)
Move validate_identifier() and validate_qualified_identifier() from
client/mod.rs and bulk.rs to a shared crate::validation module. Both
call sites now use the shared implementation, eliminating duplication.

This prepares for stored procedure support, which needs
validate_qualified_identifier() for procedure name validation (same
security requirement as savepoint and bulk insert identifier validation).
